requirements.txt
langchain
langchain-openai
langchain-community
pydantic
pyyaml
python-dotenv
chromadb
tiktoken
streamlit
